renamed to bpk-component-calendar-scrollable
